requirements.txt
streamlit
pandas
openai
matplotlib
python-docx
PyPDF2
openpyxl